+++ tests/testthat/test-conf_handle_inherited.R
@@ -1,0 +1,3 @@
+test_that("conf_handle_inherited() works", {
+  warning("TODO-20191024-7: Write test for `conf_handle_inherited()`")
+})