+++ tests/testthat/test-conf_has_inherited.R
@@ -1,0 +1,3 @@
+test_that("conf_has_inherited() works", {
+  warning("TODO-20191024-4: Write test for `conf_has_inherited()`")
+})