+++ tests/testthat/test-conf_load.R
@@ -1,17 +1,17 @@
 file_config_yml <- test_path("config.yml")
 file_config_2_yml <- test_path("config_2.yml")
 
-test_that("load() works (before)", {
+test_that("conf_load() works (before)", {
   res <- getOption("config.yml")
   expect_true(is.null(res))
 })
 
-test_that("load() works (call)", {
-  res <- load(dir = test_path())
+test_that("conf_load() works (call)", {
+  res <- conf_load(dir = test_path())
   expect_equal(res, c(file_config_yml, file_config_2_yml))
 })
 
-test_that("load() works (after)", {
+test_that("conf_load() works (after)", {
   res <- getOption("config.yml")
   expect_true(!is.null(res))
 })