+++ tests/testthat/test-conf_merge_inherited.R
@@ -1,0 +1,3 @@
+test_that("conf_merge_inherited() works", {
+  warning("TODO-20191024-6: Write test for `conf_merge_inherited()`")
+})