+++ tests/testthat/test-conf_resolve_inherited.R
@@ -1,0 +1,3 @@
+test_that("conf_resolve_inherited() works", {
+  warning("TODO-20191024-5: Write test for `conf_resolve_inherited()`")
+})