+++ tests/testthat/test-handle_inherited.R
@@ -1,3 +1,0 @@
-test_that("handle_inherited() works", {
-  warning("TODO-20191024-7: Write test for `handle_inherited()`")
-})