+++ tests/testthat/test-has_inherited.R
@@ -1,3 +1,0 @@
-test_that("has_inherited() works", {
-  warning("TODO-20191024-4: Write test for `has_inherited()`")
-})