+++ tests/testthat/test-merge_inherited.R
@@ -1,3 +1,0 @@
-test_that("merge_inherited() works", {
-  warning("TODO-20191024-6: Write test for `merge_inherited()`")
-})