+++ tests/testthat/test-resolve_inherited.R
@@ -1,3 +1,0 @@
-test_that("resolve_inherited() works", {
-  warning("TODO-20191024-5: Write test for `resolve_inherited()`")
-})